requirements.txt
python-dotenv
atlassian-python-api
dataclasses
typing
requests
markdown
